refactor: 검색 결과 컴포넌트 분리 후 페이지 서버 컴포넌트로 변경
